+++ vulp/__init__.py
@@ -6,4 +6,4 @@
 
 """Real-time motion control for Python."""
 
-__version__ = "2.2.1"
+__version__ = "2.2.2"