+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [2.2.2] - 2024-03-28
+
 ### Added
 
 - BulletInterface: Set a default servo temperature of 20 °C
@@ -181,7 +183,8 @@ All notable changes to this project will be documented in this file.
 
 First release of the project.
 
-[unreleased]: https://github.com/upkie/vulp/compare/v2.2.1...HEAD
+[unreleased]: https://github.com/upkie/vulp/compare/v2.2.2...HEAD
+[2.2.2]: https://github.com/upkie/vulp/compare/v2.2.1...v2.2.2
 [2.2.1]: https://github.com/upkie/vulp/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/upkie/vulp/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/upkie/vulp/compare/v2.0.0...v2.1.0